Added support for specifying a compiler. Specifically added clang support with warnings and optimization flags. Unfortunately, even clang 3.2 cannot yet compile TM, but when it can, I'll be ready now. Fixed a few bugs found by clang warnings.

+++ src/collection.hpp
@@ -30,7 +30,7 @@ namespace technicalmachine {
 
 class InvalidCollectionIndex : public std::out_of_range {
 	public:
-		InvalidCollectionIndex (unsigned index, unsigned size, std::string const & name):
+		InvalidCollectionIndex (unsigned index, size_t size, std::string const & name):
 			out_of_range ("Attempted to access element " + std::to_string (index) + " in a container of size " + std::to_string (size) + " with elements of type " + name + "\n")
 			{
 		}
